+++ CHANGELOG.md
@@ -5,10 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),  
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.4.1]
+
+`Active & Working`
+
+### Fixed
+  - Fixed an issue where python returns `Error processing item: 'YTMusicSearcher' object has no attribute 'get_audio_url_fast'`.
+  - Fixed `_album_art_cache` initialization.
+
 
 ## [3.4.0]
 
-`Active & Working`
+`Errors Found`
 
 ### Added